Setup babel transform for tsdown

+++ packages/code-infra/src/bundlers/tsdown.mjs
@@ -1,8 +1,11 @@
 /* eslint-disable no-console */
+import { findWorkspaceDir } from '@pnpm/find-workspace-dir';
+import { $ } from 'execa';
 import * as fs from 'node:fs/promises';
 import { builtinModules } from 'node:module';
+import * as path from 'node:path';
+import * as semver from 'semver';
 import { build as tsdown } from 'tsdown';
-import reactPlugin from '@vitejs/plugin-react';
 import {
   generateEntriesFromExports,
   getTsConfigPath,
@@ -10,6 +13,7 @@ import {
   writePkgJson,
 } from '../utils/build.mjs';
 import { copyFiles } from '../utils/copyFiles.mjs';
+import tsdownBabel from './tsdown-babel.mjs';
 
 /**
  * @TODOs -
@@ -25,16 +29,16 @@ import { copyFiles } from '../utils/copyFiles.mjs';
  */
 
 /**
- * @typedef {import('../cli/cmdBuildNew.mjs').PackageJson} PackageJson
- */
-
-/**
  * @param {Args} args
- * @param {PackageJson} pkgJson
+ * @param {import('../cli/packageJson').PackageJson} pkgJson
  * @returns {Promise<void>}
  */
 export async function build(args, pkgJson) {
   const cwd = process.cwd();
+  const workspaceDir = await findWorkspaceDir(cwd);
+  if (!workspaceDir) {
+    throw new Error('Could not find workspace root');
+  }
 
   const outDir = /** @type {any} */ (pkgJson.publishConfig)?.directory;
   await fs.rm(outDir, {
@@ -43,8 +47,8 @@ export async function build(args, pkgJson) {
   });
 
   const [exportEntries, nullEntries, binEntries] = await generateEntriesFromExports(
-    pkgJson.exports ?? {},
-    pkgJson.bin ?? {},
+    /** @type {import('../utils/build.mjs').Exports} */ (pkgJson.exports) ?? {},
+    /** @type {string | Record<string, string>} */ (pkgJson.bin) ?? {},
     { cwd },
   );
   const externals = new Set([
@@ -74,6 +78,7 @@ export async function build(args, pkgJson) {
    * @type {import('tsdown').InlineConfig}
    */
   const baseOptions = {
+    shims: true,
     watch: false,
     config: false,
     outDir,
@@ -99,16 +104,59 @@ export async function build(args, pkgJson) {
     minify: 'dce-only',
     hash: false,
     name: pkgJson.name,
-    plugins: [
-      reactPlugin({
-        jsxRuntime: 'automatic',
-        babel: {
-          babelrc: true,
-          configFile: true,
-        },
-      }),
-    ],
+    plugins: [],
   };
+
+  const babelConfigMjs = await fs
+    .stat(path.join(workspaceDir, 'babel.config.mjs'))
+    .then((stat) => stat.isFile())
+    .catch(() => false);
+  const babelConfigJs = await fs
+    .stat(path.join(workspaceDir, 'babel.config.js'))
+    .then((stat) => stat.isFile())
+    .catch(() => false);
+
+  if (babelConfigMjs || babelConfigJs) {
+    let babelRuntimeVersion = pkgJson.dependencies ? '@babel/runtime' in pkgJson.dependencies : '';
+    if (babelRuntimeVersion === 'catalog:') {
+      // resolve the version from the given package
+      // outputs the pnpm-workspace.yaml config as json
+      const { stdout: configStdout } = await $`pnpm config list --json`;
+      const pnpmWorkspaceConfig = JSON.parse(configStdout);
+      babelRuntimeVersion = pnpmWorkspaceConfig.catalog['@babel/runtime'];
+    }
+
+    if (!babelRuntimeVersion) {
+      throw new Error(
+        'package.json needs to have a dependency on `@babel/runtime` when building with `@babel/plugin-transform-runtime`.',
+      );
+    }
+    let reactVersion = '';
+    if (args.enableReactCompiler) {
+      reactVersion = semver.minVersion(pkgJson.peerDependencies?.react || '')?.version ?? 'latest';
+      const mode = process.env.MUI_REACT_COMPILER_MODE ?? 'opt-in';
+      console.log(
+        `[feature] Building with React compiler enabled. The compiler mode is "${mode}" right now.${mode === 'opt-in' ? ' Use explicit "use memo" directives in your components to enable the React compiler for them.' : ''}`,
+      );
+    }
+
+    // @ts-expect-error The plugins field exists and is an array
+    (baseOptions.plugins ?? []).push(
+      tsdownBabel({
+        root: workspaceDir,
+        optimizeClsx:
+          pkgJson.dependencies &&
+          ('clsx' in pkgJson.dependencies || 'classnames' in pkgJson.dependencies),
+        removePropTypes: pkgJson.dependencies && 'prop-types' in pkgJson.dependencies,
+        reactCompiler:
+          args.enableReactCompiler && reactVersion
+            ? {
+                reactVersion: reactVersion || 'latest',
+              }
+            : undefined,
+      }),
+    );
+  }
 
   /**
    * @type {Promise<import('tsdown').TsdownBundle[]>[]}
@@ -186,7 +234,7 @@ export async function build(args, pkgJson) {
               name: 'bin-shebang',
               // eslint-disable-next-line consistent-return
               renderChunk(code, chunk) {
-                if (chunk.isEntry && !code.startsWith('#!')) {
+                if (chunk.isEntry && !code.includes('#!/usr/bin/env')) {
                   return `#!/usr/bin/env node\n${code}`;
                 }
               },
